Migrate native to moved common lowers
