+++ gradle.properties
@@ -20,9 +20,9 @@ remoteRoot=konan_tests
 #kotlinCompilerModule=org.jetbrains.kotlin:kotlin-compiler:1.1-SNAPSHOT
 # Download artifacts of https://teamcity.jetbrains.com/viewType.html?buildTypeId=bt345
 testDataVersion=1067176:id
-kotlinCompilerRepo=http://dl.bintray.com/jetbrains/kotlin-native-dependencies
-#kotlinCompilerRepo=http://oss.sonatype.org/content/repositories/snapshots
+#kotlinCompilerRepo=http://dl.bintray.com/jetbrains/kotlin-native-dependencies
+kotlinCompilerRepo=http://oss.sonatype.org/content/repositories/snapshots
 #kotlinCompilerRepo=http://dl.bintray.com/kotlin/kotlin-dev
-kotlinCompilerModule=org.jetbrains.kotlin:kotlin-compiler:1.1-20170804.182335-746
+kotlinCompilerModule=org.jetbrains.kotlin:kotlin-compiler:1.1-20170809.024631-765
 konanVersion=0.3
 org.gradle.jvmargs='-Dfile.encoding=UTF-8'
